+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- chore(http): integrate @harborclient/http package into monorepo. (`2fdd858c`)
 - chore(sdk): integrate SDK into monorepo and update related configurations. (`f9143029`)
 - chore(deps): update @harborclient/sdk to version 1.2.5 and enhance script error handling. (`dd2ab257`)